cvpr-research-trends-2024-2025/requirements.txt
beautifulsoup4>=4.12
pandas>=2.2
pyyaml>=6.0
requests>=2.32
